Revert "Add vector Projection built-in function"
