+++ mathics/builtin/arithmetic.py
@@ -7,9 +7,6 @@ Mathematical Functions
 Basic arithmetic functions, including complex number arithmetic.
 """
 
-# This tells documentation how to sort this module
-sort_order = "mathics.builtin.mathematical-functions"
-
 
 import sympy
 import mpmath
@@ -412,8 +409,8 @@ class Im(SympyFunction):
 class Conjugate(_MPMathFunction):
     """
     <dl>
-      <dt>'Conjugate[$z$]'
-      <dd>returns the complex conjugate of the complex number $z$.
+    <dt>'Conjugate[$z$]'
+        <dd>returns the complex conjugate of the complex number $z$.
     </dl>
 
     >> Conjugate[3 + 4 I]
