+++ mathics/builtin/comparison.py
@@ -7,9 +7,6 @@ There are a number of functions for testing Expressions.
 Functions that "ask a question" have names that end in "Q". They return 'True' for an explicit answer, and 'False' otherwise.
 """
 
-# This tells documentation how to sort this module
-sort_order = "mathics.builtin.testing-expressions"
-
 from typing import Optional, Any
 
 import sympy
