+++ mathics/builtin/files_io/__init__.py
@@ -1,6 +1,3 @@
 """
 Input/Output, Files, and Filesystem
 """
-
-# This tells documentation how to sort this module
-sort_order = "mathics.builtin.input-output-files-and-filesystem"
