+++ mathics/builtin/graphics.py
@@ -5,8 +5,6 @@
 Drawing Graphics
 """
 
-# This tells documentation how to sort this module
-sort_order = "mathics.builtin.drawing-graphics"
 
 from math import sqrt
