+++ mathics/builtin/numbers/__init__.py
@@ -1,6 +1,3 @@
 """
 Integer and Number-Theoretical Functions
 """
-
-# This tells documentation how to sort this module
-sort_order = "mathics.builtin.integer-and-number-theoretical-functions"
