+++ mathics/builtin/system.py
@@ -4,8 +4,6 @@
 Global System Information
 """
 
-sort_order = "mathics.builtin.global-system-information"
-
 import gc
 import os
 import platform
