+++ mathics/core/systemsymbols.py
@@ -28,7 +28,6 @@ SymbolCondition = Symbol("Condition")
 SymbolConditionalExpression = Symbol("ConditionalExpression")
 Symbol_Context = Symbol("$Context")
 Symbol_ContextPath = Symbol("$ContextPath")
-SymbolConjugate = Symbol("Conjugate")
 SymbolContinue = Symbol("Continue")
 SymbolC = Symbol("C")
 SymbolCos = Symbol("Cos")
